% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPairwiseAlign.R
\name{AllPairwiseAlign}
\alias{AllPairwiseAlign}
\title{Compute all pairwise (global) alignments with VSEARCH}
\usage{
AllPairwiseAlign(file, cores = 1, mask = "none",
  out.name = "AllPairwiseAlign", output = NULL)
}
\arguments{
\item{file}{path to fasta file storing sequences for which all possible pairwise alignments shall be computed.}

\item{cores}{number of cores that shall be used for parallel computations.}

\item{mask}{shall the aligned sequences be maksed? Options are: \code{mask = "none"}, \code{mask = "dust"}, \code{mask = "soft"}.}

\item{out.name}{name of the output files (\code{*.uc}, \code{*_userout.txt}, \code{*.blast6out}), \code{*_fasta_pairs.fasta}, \code{*.sam}.}

\item{output}{path to a folder in which output shall be stored.}
}
\value{
A folder named \code{vsearch_pairwise_align} will be created and the following files will be stored in this output folder:

\itemize{
\item \code{*.uc} USEARCH cluster format generated by VSEARCH storing the sequence cluster information of aligned sequences.
\item \code{*_userout.txt} a tab separated file storing the alignment information in the columns: \code{query}, \code{target}, \code{sequence identity (ID)}.
\item \code{*.blast6out} BLAST output format of the pairwise (global) sequence alignments generated by VSEARCH.
\item \code{*_fasta_pairs.fasta} fasta file storing the alignments.
\item \code{*.sam} SAM file storing the alignments.
}
}
\description{
This function is a wrapper function to compute all pairwise (global) alignments using VSEARCH.
}
\details{
To be able to use this function the VSEARCH command line tool needs to be installed.
}
\references{
https://github.com/torognes/vsearch
}
\author{
Hajk-Georg Drost
}
