% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.short.sci.name.R
\name{assign.short.sci.name}
\alias{assign.short.sci.name}
\title{Transform long file names to short scientific names}
\usage{
assign.short.sci.name(data)
}
\arguments{
\item{data}{a \code{tibble} from \code{\link{meta.seq.space}}.}
}
\description{
This function takes a \code{tibble} as input
and transform long file names to short scientific names.
}
\author{
Hajk-Georg Drost
}
