% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcolor.R
\name{bcolor}
\alias{bcolor}
\title{Beautiful colors for plots}
\usage{
bcolor(n)
}
\arguments{
\item{n}{number of different colors that shall be returned.}
}
\description{
A distinguished color scheme for scientific plots.
}
\examples{
# return 5 colors 
bcolor(5)
}
\author{
Hajk-Georg Drost
}

