% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoMapChromosomeFile.R
\name{chromoMapChromosomeFile}
\alias{chromoMapChromosomeFile}
\title{Generate chromsome file for chromoMap package visualization}
\usage{
chromoMapChromosomeFile(data, output)
}
\arguments{
\item{data}{a \code{\link{data.frame}} with the following column format:
\itemize{
\item column 1: \code{data$chromosome_name}
\item column 2: \code{data$chromosome_start}
\item column 3: \code{data$chromosome_end}
\item column 4: \code{data$centromere_start}
}}

\item{output}{a file path (file name) to the output file.}
}
\description{
Generate chromsome file for chromoMap package visualization.
}
\references{
https://cran.r-project.org/web/packages/chromoMap/vignettes/chromoMap.html
}
\author{
Hajk-Georg Drost
}
