% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRepBase.R
\name{CleanRepBase}
\alias{CleanRepBase}
\title{Clean the initial Repbase database for BLAST}
\usage{
CleanRepBase(repbase.file, output.file)
}
\arguments{
\item{repbase.file}{fasta file storing the corresponding Repbase annotation, e.g. \code{athrep.ref}.}

\item{output.file}{name/path of the cleaned Repbase annotation file.}
}
\description{
This function cleans the headers of the Repbase fasta files
so that they can be used to create a blast-able database.
}
\details{
The Repbase database can be downloaded after registration at http://www.girinst.org/repbase/.
The corresponding files as they are however, cannot be converted into a blast-able database.
Hence, a pre-filtering step is neccessary to be able to use this database with the e.g. \code{\link{QueryRepBase}}
function.
}
\examples{
\dontrun{
# PreProcess Repbase: A thaliana
# and save the output into the file "Athaliana_repbase.ref"
CleanRepBase(repbase.file = "athrep.ref",
             output.file  = "Athaliana_repbase.ref")

}
}
\author{
Hajk-Georg Drost
}
\references{
http://www.girinst.org/repbase/
}

