% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust2fasta.R
\name{clust2fasta}
\alias{clust2fasta}
\title{Export sequences of TEs belonging to the same cluster to fasta files}
\usage{
clust2fasta(LTRpred.tbl, seqs)
}
\arguments{
\item{LTRpred.tbl}{\code{data.frame} generated by \code{\link{LTRpred}}.}

\item{seqs}{path to the fasta file storing the TE sequences.}
}
\description{
Export sequences of TEs belonging to the same cluster to separate fasta files.
Each cluster will be stored in an individual fasta file.
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{cluster.members}}
}

