% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.members.R
\name{cluster.members}
\alias{cluster.members}
\title{Select members of a specific cluster}
\usage{
cluster.members(LTRpred.tbl, cluster = NULL)
}
\arguments{
\item{LTRpred.tbl}{\code{data.frame} generated by \code{\link{LTRpred}}.}

\item{cluster}{name of the cluster that shall be searched for.}
}
\details{
This function allows users to select all members of a specific cluster.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
# read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)

# extract memebers of cluster 'cl_108'
cluster.members(pred, cluster = "cl_108")

# or safe the sequences of the cluster as fasta
pred2fasta(cluster.members(pred, cluster = "cl_108"), "cl_108_seqs.fa")
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{clust2fasta}}
}

