% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUSTpred.R
\name{CLUSTpred}
\alias{CLUSTpred}
\title{Cluster Sequences with VSEARCH}
\usage{
CLUSTpred(file, similarity = 0.9, strand = "both", cores = 1,
  out.name = "CLUSTpred", output = NULL)
}
\arguments{
\item{file}{path to predicted LTR transposons generated by \code{\link{LTRpred}} (in fasta format).}

\item{similarity}{reject if sequence similarity is lower than this threshold.}

\item{strand}{cluster using plus or both strands.}

\item{cores}{number of cores that shall be used for parallel computations.}

\item{out.name}{name of the output files (\code{*.uc}, \code{*.log}, \code{*.blast6out}).}

\item{output}{path to a folder in which output shall be stored.}
}
\value{
First the following files generated by VSEARCH are stored in the output folder (default: \code{\link{getwd}}):

\itemize{
\item \code{*.uc} USEARCH cluster format generated by VSEARCH storing the sequence cluster information of the input LTR transposons.
\item \code{*.log} a log file of the VSEARCH run.
\item \code{*.blast6out} BLAST output generated by VSEARCH storing the BLAST hit information of the input LTR transposons.
}

  
A USEARCH cluster format (*.uc file extension) table (see \code{\link{read.uc}} for specifications).
}
\description{
Cluster putative LTR transposons predicted by \code{\link{LTRpred}} using VSEARCH.
}
\details{
To be able to use this function the VSEARCH command line tool needs to be installed.
}
\references{
https://github.com/torognes/vsearch
}
\author{
Hajk-Georg Drost
}
