% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn2bed.R
\name{cn2bed}
\alias{cn2bed}
\title{Write copy number estimation results to BED file format.}
\usage{
cn2bed(cn.pred, type = "solo", filename = "copy_number_est", sep = "\\t",
  output = NULL)
}
\arguments{
\item{cn.pred}{\code{data.frame} object returned by \code{\link{ltr.cn}} (if \code{type = "solo"}).}

\item{type}{type of copy number estimation: \code{\link{ltr.cn}} (if \code{type = "solo"}).}

\item{filename}{name of the output file (will be extended by "*.csv").}

\item{sep}{column separator.}

\item{output}{path in which the output file shall be stored.}
}
\description{
Format copy number estimation output to BED file format.
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{ltr.cn}}, \code{\link{LTRpred}}
}

