% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePreds.R
\name{combinePreds}
\alias{combinePreds}
\title{Combine LTRpred prediction files}
\usage{
combinePreds(folder, quality.filter = FALSE, sim = 70, n.orfs = 0)
}
\arguments{
\item{folder}{path to folder in which prediction files are stored.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not.}

\item{sim}{If \code{quality.filter = TRUE}: LTR similarity threshold. Only putative LTR transposons that fulfill this 
LTR similarity threshold will be retained.}

\item{n.orfs}{If \code{quality.filter = TRUE}: minimum number of ORFs detected in the putative LTR transposon.}
}
\description{
Given a path to a folder that stores several LTRpred prediction files,
this function imports the individual prediction files and combines them to one
large LTRpred prediction file.
}
\author{
Hajk-Georg Drost
}
