% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfam.query.R
\name{dfam.query}
\alias{dfam.query}
\title{Query the Dfam database to annotate putative LTRs}
\usage{
dfam.query(seq.file, Dfam.db = NULL, eval = 0.001, cores = 1,
  output.folder = getwd())
}
\arguments{
\item{seq.file}{file path to the putative LTR transposon sequences in \code{fasta} format.}

\item{Dfam.db}{folder path to the local Dfam database or \code{Dfam.db = "download"} in case the Dfam
database shall be automatically downloaded before performing query analyses.}

\item{eval}{E-value threshhold to perform the HMMer search against the Dfam database.}

\item{cores}{number of cores to use to perform parallel computations.}

\item{output.folder}{folder path to store the annotation results.}
}
\description{
Validate or annotate putative LTR
transposons that have been predicted using \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.
}
\details{
The Dfam database provides a collection of curated transposable element annotations.
}
\author{
Hajk-Georg Drost
}
