% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.move.R
\name{file.move}
\alias{file.move}
\title{Move folders from one location to another}
\usage{
file.move(from, to)
}
\arguments{
\item{from}{path from where to copy a folder.}

\item{to}{path where to copy the corresponding folder specified in \code{from}.}
}
\description{
A small helper function to make file handling
in R easier.
}
\author{
Hajk-Georg Drost
}

