% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.jumpers.R
\name{filter.jumpers}
\alias{filter.jumpers}
\title{Detect LTR retrotransposons that are potential jumpers}
\usage{
filter.jumpers(LTRpred.tbl, ltr.similarity = 95)
}
\arguments{
\item{LTRpred.tbl}{a \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{ltr.similarity}{LTR similarity threshold. Default is \code{ltr_similarity = 95}.}
}
\description{
This function applies specific filter criteria to
screen for LTR retrotransposons predicted by \code{\link{LTRpred}}
that are potentially able to transpose due to their sequence features.
}
\details{
This ...
}
\examples{
\dontrun{
# generate de novo LTR transposon prediction
pred <- LTRpred(genome.file = "TAIR10_chr_all.fas",
                trnas       = "plantRNA_Arabidopsis.fsa",
                hmms        = "hmm_*")
                
# detect potential jumpers               
filter.jumpers(pred)
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}
}

