% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.uc.R
\name{filter.uc}
\alias{filter.uc}
\title{Filter for cluster members}
\usage{
filter.uc(cluster.file)
}
\arguments{
\item{cluster.file}{a \code{data.frame} \code{*.uc} (USEARCH cluster) format (imported with \code{\link{read.uc}}).}
}
\description{
Filter for cluster memebers in a \code{*.uc} format table.
}
\author{
Hajk-Georg Drost
}
