% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pred.filenames.R
\name{get.pred.filenames}
\alias{get.pred.filenames}
\title{Retrieve file names of files genereated by LTRpred}
\usage{
get.pred.filenames(LTRpred.folder, ws.wrap = FALSE)
}
\arguments{
\item{LTRpred.folder}{path to a \code{\link{LTRpred}} generated folder.}

\item{ws.wrap}{shall white space separated path names be wrapped with \code{'file name'}.}
}
\description{
Useful helper function to generate automated file names
when crawling LTRpred generated result folders.
}
\examples{
# without wrapped whitespaces
get.pred.filenames("path/to/LTRpred.folder")

# with wrapped whitespaces
get.pred.filenames("path to/LTRpred.folder", ws.wrap = TRUE )
}
\author{
Hajk-Georg Drost
}
