% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seqs.R
\name{get.seqs}
\alias{get.seqs}
\title{Quickly retrieve the sequences of a \code{Biostrings} object}
\usage{
get.seqs(object, type = "DNA")
}
\arguments{
\item{object}{a \code{Biostrings} object.}

\item{type}{either \code{type = "DNA"} or \code{type = "AA"}.}
}
\description{
Helper function to retrieve the sequences of a \code{Biostrings} object.
}
\examples{
# read example sequences
seqs <- system.file("nt.fa",package = "LTRpred")
input.seqs <- read.seqs(seqs)

retrieve sequences of Biostrings object
head(get.seqs(input.seqs))
}
\author{
Hajk-Georg Drost
}

