% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPlotAssemblyVersions.R
\name{gridPlotAssemblyVersions}
\alias{gridPlotAssemblyVersions}
\title{Plot retrotransposon age distributions for predictions of different genome assembly versions}
\usage{
gridPlotAssemblyVersions(
  data,
  quality.filter = FALSE,
  text.size = 18,
  xlab = "LTR similarity",
  ylab = "LTR retrotransposon count",
  main.text = "",
  y.ticks = 6,
  sim = 70,
  n.orfs = 0
)
}
\arguments{
\item{data}{prediction files returned by \code{\link[LTRpred]{LTRpred}}. Usually a combined prediction file that was generated with \code{\link{combinePreds}}.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not.}

\item{text.size}{size of x-axis, y-axis, and title text.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main.text}{title text.}

\item{y.ticks}{number of ticks that shall be drawn on the y-axis.}

\item{sim}{If \code{quality.filter = TRUE}: LTR similarity threshold. Only putative LTR transposons that fulfill this 
LTR similarity threshold will be retained.}

\item{n.orfs}{If \code{quality.filter = TRUE}: minimum number of ORFs detected in the putative LTR transposon.}
}
\description{
Compare the abundance of retrotransposon predictions between different genome assembly versions. This way a genome assembly version bias can be either ruled out
or considered for subsequent conclusions.
}
\author{
Hajk-Georg Drost
}
