% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.with.genome.tbl.R.R
\name{join.with.genome.tbl}
\alias{join.with.genome.tbl}
\title{Join \code{gm_files} returned by \code{generate.multi.quality.filter.meta} with a genome information table}
\usage{
join.with.genome.tbl(meta.summary.file, genome.tbl)
}
\arguments{
\item{meta.summary.file}{a meta.summary.file.}

\item{genome.tbl}{a genome.tbl.}
}
\description{
Join \code{gm_files} returned by \code{generate.multi.quality.filter.meta} with a genome information table.
}
\author{
Hajk-Georg Drost
}
