% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRdigest.R
\name{LTRdigest}
\alias{LTRdigest}
\title{Run LTRdigest to predict putative LTR Retrotransposons}
\usage{
LTRdigest(input.gff3, genome.file, aaout = "yes", aliout = "yes",
  pptlen = c(8, 30), uboxlen = c(3, 30), pptradius = 30, trnas = NULL,
  pbsalilen = c(11, 30), pbsoffset = c(0, 5), pbstrnaoffset = c(0, 5),
  pbsmaxedist = 1, pbsradius = 30, hmms = NULL, pdomevalcutoff = 1e-05,
  pbsmatchscore = 5, pbsmismatchscore = -10, pbsinsertionscore = -20,
  pbsdeletionscore = -20, pfam.ids = NULL, cores = 1, index.file = NULL,
  output.path = NULL)
}
\description{
This function implements an interface between R and
the LTRdigest command line tool to predict putative LTR retrotransposons from R.
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}

