% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRpred.meta.R
\name{LTRpred.meta}
\alias{LTRpred.meta}
\title{Perform Meta-Analyses with LTRpred}
\usage{
LTRpred.meta(genome.folder, result.folder = getwd(), ...)
}
\arguments{
\item{genome.folder}{path to the folder storing all unmasked genomes for which \code{\link{LTRpred}}
based \code{de novo} LTR retrotransposon prediction shall be performed.}

\item{result.folder}{folder in which \code{LTRpred} results shall be stored.}

\item{...}{all parameters of \code{\link{LTRpred}}.}
}
\description{
Run \code{\link{LTRpred}} on several genomes (sequentially) that are stored in a given folder.
}
\details{
This function provides a crawler to run \code{\link{LTRpred}} sequencially
on any number of genomes stored within the same folder.
The result will be saved on in the \code{result.folder} and can be analysed with the
\code{\link{meta.apply}} function.
}
\examples{
\dontrun{
# perform a meta analysis on multiple genomes
# stored in a genomes folder
LTRpred.meta(genome.folder = "Genomes/",
             result.folder = "LTRpredResults"
             trnas         = "plantRNA_Arabidopsis.fsa",
             hmms          = "hmm_*")
}
}
\author{
Hajk-Georg Drost
}

