% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchPredictionWithAnnotation.R
\name{MatchPredictionWithAnnotation}
\alias{MatchPredictionWithAnnotation}
\title{Match LTRharvest or LTRdigest prediction with a given Annotation file in GFF3 format}
\usage{
MatchPredictionWithAnnotation(ltr.digest.prediction, annotation.file,
  strand.ori = "+", overlap.type = "any")
}
\arguments{
\item{ltr.digest.prediction}{a \code{data.frame} returned by \code{\link{LTRharvest}} or \code{\link{LTRdigest}}.}

\item{annotation.file}{an genome annotation file in gff3 format.}

\item{strand.ori}{the strand orientation that shall be matched between the the prediction and the annotation.
Please note that some predictions do not include strand information. In case strand information should not be considered
please specify \code{strand.ori = NULL}.}

\item{overlap.type}{type of overlap between the LTR transposon prediction and the genomic feature annotated in the annotation file. 
Inherited from \pkg{IRanges}. Types can be \code{overlap.type = "any"}, \code{overlap.type = "start"}, \code{overlap.type = "end"},
\code{overlap.type = "within"}, or \code{overlap.type = "equal"}.}
}
\description{
This function allows users to match LTRharvest or LTRdigest predictions with a given Anootation file in GFF3 format
to check whether or not predicted LTR transposons overlap with annotated genomic features (in this case with genes).
}
\examples{
\dontrun{
#' ## Hypothetical Example for Arabidopsis thaliana
# Generate LTR transposon prediction for A. thaliana
 LTRharvest("Genome/TAIR10_chr_all.fas")

 LTRdigest(input.gff3        = "TAIR10_chr_all/TAIR10_chr_all_Prediction.gff", 
           genome.file       = "Genome/TAIR10_chr_all.fas",
           trnas             = "araTha1-tRNAs.fa",
           hmms              = "hmm_*",
           cores             = 1)

# Read the output of LTRdigest()
LTRdigest <- read.prediction(gff.file    = "TAIR10_chr_all_LTRdigestPrediction.gff",
                             tabout.file = "TAIR10_chr_all-ltrdigest_tabout.csv",
                             program     = "LTRdigest")
                                            
# Match predicted LTR transposons with annotation file (feature = genes)        
Anno <- MatchPredictionWithAnnotation(
              ltr.digest.prediction = Ath.LTRdigest.prediction$ltr.retrotransposon,
               annotation.file       = "Arabidopsis_thaliana.TAIR10.30.chr.gff3")
}
}
\author{
Hajk-Georg Drost
}

