% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.apply.R
\name{meta.apply}
\alias{meta.apply}
\title{Apply functions to meta data generated by \code{LTRpred}}
\usage{
meta.apply(result.folder, FUN)
}
\arguments{
\item{result.folder}{\code{result.folder} generated by \code{\link{LTRpred.meta}}.}

\item{FUN}{function to apply to meta data returned by \code{\link{LTRpred.meta}}.}
}
\description{
Apply any function to the results generated by meta analyses
}
\seealso{
\code{\link{LTRpred.meta}}, \code{\link{LTRpred}}
}
\author{
Hajk-Georg Drost
}
