% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.seq.space.R
\name{meta.seq.space}
\alias{meta.seq.space}
\title{Quantification of the repeat space for multiple \code{Repeat Masker} files}
\usage{
meta.seq.space(rm.folder, type = "rm", repeat.type = "all",
  rename = FALSE, save.file = TRUE)
}
\arguments{
\item{rm.folder}{file path to the \code{Repeat Masker} files.}

\item{type}{type of input data. Option are:
\itemize{
 \item \code{type = "rm"} - \code{Repeat Masker} files.
}}

\item{repeat.type}{type of element for which total repeat space shall be quantified.
Options are:

\itemize{
 \item \code{repeat.type = "all"} include all types of repeats.
 \item \code{repeat.type = "LTR"} include only LTR retrotransposons.
}}

\item{rename}{logical value indicating whether or not long file names shall be transformed to short scientific names.}

\item{save.file}{logical value indicating whether or not the result \code{tibble} shall be stored locally.}
}
\description{
The repeat space for multiple \code{Repeat Masker} files is quantified and
stored in one \code{tibble}.
}
\author{
Hajk-Georg Drost
}
