% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.summarize.R
\name{meta.summarize}
\alias{meta.summarize}
\title{Summarize (concatenate) all predictions of a \code{LTRpred.meta} run}
\usage{
meta.summarize(
  result.folder,
  ltr.similarity = 70,
  quality.filter = TRUE,
  n.orfs = 0,
  strategy = "default"
)
}
\arguments{
\item{result.folder}{path to meta result folder generated by \code{\link{LTRpred.meta}}.}

\item{ltr.similarity}{only count elements that have an LTR similarity >= this threshold.}

\item{quality.filter}{optimize search to remove potential false positives (e.g. duplicated genes, etc.). See \code{Details} for further information on the filter criteria.}

\item{n.orfs}{minimum number of Open Reading Frames that must be found between the LTRs (if \code{quality.filter = TRUE}). See \code{Details} for further information on quality control.}

\item{strategy}{quality filter strategy. Options are
\itemize{
\item \code{strategy = "default"} : see section \code{Quality Control} 
\item \code{strategy = "stringent"} : in addition to filter criteria specified in section \code{Quality Control},
the filter criteria \code{!is.na(protein_domain)) | (dfam_target_name != "unknown")} is applied
}}
}
\value{
a \code{LTRpred.tbl} storing the \code{\link{LTRpred}} prediction \code{data.frames} for all species in the meta result folder generated by \code{\link{LTRpred.meta}}.
}
\description{
Crawl through all genome predictions performed with \code{\link{LTRpred.meta}}
and concatenate the prediction files for each species in the meta result folder
generated by \code{\link{LTRpred.meta}} to a meta-species \code{data.frame}.
}
\details{
This function crawls through each genome stored in the meta result folder
generated by \code{\link{LTRpred.meta}} and performs the following procedures:

\itemize{
\item \strong{Step 1:} For each genome: Read the \code{*._LTRpred_DataSheet.csv} file generated by \code{\link{LTRpred}}.
\item \strong{Step 2:} For each genome: Perform quality filtering and selection of elements having at least \code{ltr.similarity} sequence similarity between their LTRs (if \code{quality.filter = TRUE}). Otherwise no quality filtering is performed.
\item \strong{Step 3:} Summarize all genome predictions in the meta-folder to one meta-species \code{data.frame}.
}

\strong{Quality Filtering}

The aim of the quality filtering step is to reduce the potential false positive
LTR transposons that were predicted by \code{\link{LTRpred}}. These false positives can be
duplicated genes, or other homologous repetitive elements that fulfill the LTR similarity
criteria, but do not have any Primer Binding Site, Open Reading Frames, Gag and Pol
proteins, etc. To reduce the number of false positives, the following filters are applied
to discard false positive LTR transposons.

\itemize{
\item \code{ltr.similarity}: Minimum similarity between LTRs. All TEs not matching this
 criteria are discarded.
 \item \code{n.orfs}: minimum number of Open Reading Frames that must be found between the
  LTRs. All TEs not matching this criteria are discarded.
 \item \code{PBS or Protein Match}: elements must either have a predicted Primer Binding
 Site or a protein match of at least one protein (Gag, Pol, Rve, ...) between their LTRs. All TEs not matching this criteria are discarded.
}
}
\author{
Hajk-Georg Drost
}
