% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.count.R
\name{motif.count}
\alias{motif.count}
\title{Low level function to detect motifs in strings}
\usage{
motif.count(seq.file, motif, as.ratio = FALSE)
}
\arguments{
\item{seq.file}{path to the genomic sequecne file of interest (e.g. LTR TE seqs predicted by
\code{\link{LTRpred}}).}

\item{motif}{a character string or vector of strings which shall be counted within each sequence.}

\item{as.ratio}{shall count values be returned as asbolute frequency (count value) or as relative frequency (percentage).}
}
\description{
Find a specific motif or a sequence of motifs within
genomic sequences.
}
\examples{
# find number of "CG" motifs in predicted LTR transposons
motif.count(seq.file = system.file("LTRseqs.fas",package = "LTRpred"), 
            motif    = "CG")
            
# find number of "CG" motifs in predicted LTR transposons: rel. frequency
motif.count(seq.file = system.file("LTRseqs.fas",package = "LTRpred"), 
            motif    = "CG",
            as.ratio = TRUE)             
}
\author{
Hajk-Georg Drost
}

