% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_copynumber_individual.R
\name{plot_copynumber_individual}
\alias{plot_copynumber_individual}
\title{Plot the median cluster copy number distribution of putative LTR transposons for and individual species}
\usage{
plot_copynumber_individual(
  data,
  element.type = "full_retrotransposon",
  plot.type = "boxplot",
  similarity.bin = 2,
  min.sim = 70,
  quality.filter = FALSE,
  n.orfs = 0,
  include.unique = FALSE,
  xlab = "LTR \% Similarity",
  ylab = "Copy Number",
  main = "",
  legend.title = "Similarity between LTRs",
  arrow_lab = FALSE,
  y.ticks = 10
)
}
\arguments{
\item{data}{the \code{\link{data.frame}} generated by \code{\link[LTRpred]{LTRpred}}.}

\item{element.type}{which width shall be visualized. Full retrotransposon width or only LTR element cluster copy number: 
choose \code{element.type = "full_retrotransposon"} or \code{element.type = "ltr_element"}. 
In case \code{element.type = "ltr_element"} is chosen then the width of the left LTR element is 
used to visualize width vs. seq. similarity between the LTRs.}

\item{plot.type}{type of visualization: either \code{plot.type = "boxplot"} or \code{plot.type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. See Description for details.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{include.unique}{shall unique elements be visualized together with the clusters?}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{title label.}

\item{legend.title}{title of legend.}

\item{arrow_lab}{whether or not data points shall be labelled with arrows (ggrepel).}

\item{y.ticks}{number of ticks at the y-axis. Default is \code{y.ticks = 10}.}
}
\description{
Visualize the correlation between LTR transposon age
(measured in LTR similarity) and the median cluster copy number of the corresponding LTR retrotransposon (the complete sequence) 
or the width of the left LTR element for an individual species stored in the input \code{data}.
}
\details{
This function visualizes the correlation between LTR transposon age
(measured in LTR similarity) and the cluster copy number of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element. Using this visualization approach, different classes of LTR retrotransposons can be detected due to their width and age correlation.
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}
