% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltrwidth_individual.R
\name{plot_ltrwidth_individual}
\alias{plot_ltrwidth_individual}
\title{Plot the width distribution of putative LTR transposons}
\usage{
plot_ltrwidth_individual(
  data,
  element.type = "full_retrotransposon",
  plot.type = "boxplot",
  similarity.bin = 2,
  min.sim = 70,
  quality.filter = FALSE,
  n.orfs = 0,
  xlab = "LTR \% Similarity",
  ylab = "LTR Retrotransposon length in bp",
  main = "",
  legend.title = "Similarity between LTRs",
  y.ticks = 10
)
}
\arguments{
\item{data}{the \code{\link{data.frame}} generated by \code{\link[LTRpred]{LTRpred}}.}

\item{element.type}{which width shall be visualized. Full retrotransposon width or only LTR element width: 
choose \code{element.type = "full_retrotransposon"} or \code{element.type = "ltr_element"}. 
In case \code{element.type = "ltr_element"} is chosen then the width of the left LTR element is 
used to visualize width vs. seq. similarity between the LTRs.}

\item{plot.type}{type of visualization: either \code{plot.type = "boxplot"} or \code{plot.type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. See Description for details.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{title label.}

\item{legend.title}{title of legend.}

\item{y.ticks}{number of ticks at the y-axis. Default is \code{y.ticks = 10}.}
}
\description{
Visualize the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element.
}
\details{
This function visualizes the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element. Using this visualization approach, different classes of LTR retrotransposons can be detected due to their width and age correlation.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred(genome.file = "TAIR10_chr_all.fas",
                    trnas       = "plantRNA_Arabidopsis.fsa",
                    hmms        = "hmm_*")
                    
# visualize the collelation between LTR transposon age and width
# of predicted  A. thaliana LTR retrotransposons
plot_ltrwidth_individual(Ath.Pred, plot.type = "violin")

# visualize the collelation between LTR retrotransposon age and width
# of predicted  A. thaliana LTR element
plot_ltrwidth_individual(Ath.Pred, element.type = "full_retrotransposon")
}
}
\author{
Hajk-Georg Drost
}
