% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pbswidth_kingdom.R
\name{plot_pbswidth_kingdom}
\alias{plot_pbswidth_kingdom}
\title{Plot the median PBS widths for kingdoms of life}
\usage{
plot_pbswidth_kingdom(data, summary.type = "median",
  plot.type = "boxplot", similarity.bin = 2, min.sim = 70,
  quality.filter = FALSE, n.orfs = 0, xlab = "LTR \% Similarity",
  ylab = "Median PBS length in bp", main = "",
  legend.title = "Similarity between LTRs", y.ticks = 10)
}
\arguments{
\item{data}{the \code{\link{data.frame}} generated by \code{\link[LTRpred]{LTRpred}}.}

\item{summary.type}{summary statistic to quantify total LTR length or LTR element length.
Options are:
\itemize{
 \item \code{summary.type = "median"} - for quantifying the median length of all elements
}}

\item{plot.type}{type of visualization: either \code{plot.type = "boxplot"} or \code{plot.type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. See Description for details.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{title label.}

\item{legend.title}{title of legend.}

\item{y.ticks}{number of ticks at the y-axis. Default is \code{y.ticks = 10}.}
}
\description{
Visualize the correlation between LTR transposon age
(measured in LTR similarity) and the median PBS width of the corresponding LTR retrotransposon.
}
\details{
This function visualizes the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element. Using this visualization approach, different classes of LTR retrotransposons can be detected due to their width and age correlation.
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}
