% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromoMap.R
\name{plotChromoMap}
\alias{plotChromoMap}
\title{Plot a predicted retrotransposons along the chromosomes}
\usage{
plotChromoMap(pred, genome.file, centromere_starts, ...)
}
\arguments{
\item{pred}{LTRpred.tbl generated with \code{\link{LTRpred}}.}

\item{genome.file}{a file path to the genome assembly file in fasta format for which chromosomes shall be visualized.}

\item{centromere_starts}{a numeric vector containing the start coordinates of the centromers (for all chromosomes in \code{genome.file}).}
}
\description{
LTR retrotransposons predicted with \code{\link{LTRpred}}
will be visualized along the chromsome.
}
\examples{
test_genome <- system.file("Hsapiens_ChrY.fa", package = "LTRpred")
test_pred <- LTRpred::read.ltrpred(system.file("Hsapiens_ChrY_LTRpred_DataSheet.tsv", package = "LTRpred"))
test_centromere_starts <- 55000
# generate visualization
LTRpred::plotChromoMap(test_pred, test_genome, test_centromere_starts)
}
\references{
https://cran.r-project.org/web/packages/chromoMap/vignettes/chromoMap.html
}
\author{
Hajk-Georg Drost
}
