% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDomain.R
\name{plotDomain}
\alias{plotDomain}
\title{Visualize the Protein Domain distribution of predicted LTR retrotransposons}
\usage{
plotDomain(data, xlab = "\\nDetected Protein Domains",
  ylab = "Predicted LTR retrotransposons (count)\\n", main = "",
  legend.title = "Protein Domains")
}
\arguments{
\item{data}{\code{data.frame} returned by \code{\link[LTRpred]{LTRpred}}.}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{main}{title text.}

\item{legend.title}{text of the legend title.
Either \code{annotation = "Dfam"} or \code{annotation = "Repbase"}.}
}
\description{
Plot the Superfamily distribution of predicted LTR retrotransposons of
annotations generated via Dfam or Repbase query.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRlearn")
# read LTRpred generated prediction file (data sheet)
pred <- LTRpred::read.ltrpred(pred.file)

# visualize superfamily distribution
plotFamily(pred)
}
\seealso{
\code{\link[LTRpred]{LTRpred}}, \code{\link[LTRpred]{read.ltrpred}}, \code{\link[LTRpred]{dfam.query}}, \code{\link[LTRpred]{repbase.query}}
}
\author{
Hajk-Georg Drost
}
