% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFamily.R
\name{PlotFamily}
\alias{PlotFamily}
\title{Visualize the Superfamily distribution of predicted LTR retrotransposons}
\usage{
PlotFamily(LTRpred.tbl, xlab = "\\nAnnotated Superfamily Name",
  ylab = "Predicted LTR retrotransposons (count)\\n",
  main = "Superfamily distribution of predicted LTR retrotransposons",
  legend.title = "Superfamily", annotation = "Dfam")
}
\arguments{
\item{LTRpred.tbl}{\code{data.frame} returned by \code{\link{LTRpred}}.}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{main}{title text.}

\item{legend.title}{text of the legend title.}

\item{annotation}{which database has been used to generate annotation queries when running \code{\link{LTRpred}}? 
Either \code{annotation = "Dfam"} or \code{annotation = "Repbase"}.}
}
\description{
Plot the Superfamily distribution of predicted LTR retrotransposons of
annotations generated via Dfam or Repbase query.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
# read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)

# visualize superfamily distribution
PlotFamily(pred)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{read.ltrpred}}, \code{\link{dfam.query}}, \code{\link{repbase.query}}
}

