% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInterSpeciesCluster.R
\name{plotInterSpeciesCluster}
\alias{plotInterSpeciesCluster}
\title{Plot inter species similarity between TEs (for a specific cluster)}
\usage{
plotInterSpeciesCluster(
  cluster.file,
  cluster,
  xlab = "Species",
  ylab = "\% Similarity to Reference",
  legend.title = "Species"
)
}
\arguments{
\item{cluster.file}{a cluster file in \code{*.uc} format (imported with e.g. \code{\link[LTRpred]{filter.uc}} and \code{\link[LTRpred]{read.uc}}).}

\item{cluster}{numeric value denoting the cluster number of interest.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{legend.title}{legend title text.}
}
\description{
After clustering TE sequences across different species with \code{\link[LTRpred]{CLUSTpred}}
the similarity of TEs from different species within one cluster can be visualized.
}
\author{
Hajk-Georg Drost
}
