% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRAge.R
\name{PlotLTRAge}
\alias{PlotLTRAge}
\title{Plot the age distribution of predicted LTR transposons}
\usage{
PlotLTRAge(LTRpred.tbl, type = "hist", stack.fill = "protein_domain",
  similarity.bin = 2, min.sim = 70, quality.filter = TRUE, n.orfs = 1,
  xlab = "LTR \% Similarity", ylab = "Frequency",
  main = "LTR Age Distribution", legend.title = "LTR Similarity")
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{type}{type of histogram. Either normal histogram (\code{type = "hist"}) or stacked histogram (\code{type = "stack"}, see also \code{stack.fill}).
If \code{type = "stack"} is specified then}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.
Default is \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization. 
All elements not fulfilling this similarity threshold are filtered out. Default is \code{min.sim = 2}.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. See Description for details.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{legend.title}{legend text.}
}
\description{
This function visualizes the age distribution of
predicted LTR transposons generated with \code{\link{LTRpred}}.

The age pf LTR transposons is defined by the sequence similarity between it's
3' and 5' LTR. Evolutionary young (recent) LTR transposons tend to have very similar
LTRs (up to 100\% sequence similarity), whereas evolutionary older LTR transposons
tend to have less similar LTRs.
}
\details{
This way of visualizing the age distribution of LTR transposons
allows users to examine the rate of recent transposition events in extant organisms. 

LTR similarity values are binned in intervals (per default 2.5\% intervals) which can be modified
by the \code{similarity.bin} argument.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred(genome.file = "TAIR10_chr_all.fas")
# visualize the age distribution of predicted  A. thaliana LTR transposons
PlotLTRAge(Ath.Pred)
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRharvest}}, \code{\link{LTRdigest}},
\code{\link{PlotLTRWidth}}, \code{\link{PlotRanges}}
}

