% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRAgeDistribution.R
\name{PlotLTRAgeDistribution}
\alias{PlotLTRAgeDistribution}
\title{Plot the age distribution of predicted LTR transposons}
\usage{
PlotLTRAgeDistribution(ltr.harvest.prediction, similarity.bin = NULL,
  min.sim = NULL)
}
\description{
This function
}
\details{
This
}
\examples{
\dontrun{

}

}
\author{
Hajk-Georg Drost
}

