% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLTRRange.R
\name{plotLTRRange}
\alias{plotLTRRange}
\title{Plot the location of putative LTR transposons along the chromosomes}
\usage{
plotLTRRange(data)
}
\arguments{
\item{data}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}
}
\description{
Visualization of the location of putative LTR transposons along chromosomes.
}
\author{
Hajk-Georg Drost
}
