% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRTransposonWidthDistribution.R
\name{PlotLTRTransposonWidthDistribution}
\alias{PlotLTRTransposonWidthDistribution}
\title{Plot the width distribution of putative LTR transposons}
\usage{
PlotLTRTransposonWidthDistribution(ltr.harvest.prediction, type = "boxplot",
  similarity.bin = NULL, min.sim = NULL)
}
\description{
This
}
\details{
This
}
\examples{
\dontrun{

}

}
\author{
Hajk-Georg Drost
}

