% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRWidth.R
\name{PlotLTRWidth}
\alias{PlotLTRWidth}
\title{Plot the width distribution of putative LTR transposons}
\usage{
PlotLTRWidth(LTRpred.tbl, element.type = "full_retrotransposon",
  plot.type = "violin", similarity.bin = 2, min.sim = 70,
  quality.filter = TRUE, n.orfs = 1, xlab = "\\nLTR \% Similarity",
  ylab = "LTR Retrotransposon length in bp\\n",
  main = "Length Distribution", legend.title = "LTR similarity")
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{element.type}{which width shall be visualized. Full retrotransposon width or only LTR element width: choose \code{element.type = "full_retrotransposon"} or \code{element.type = "ltr_element"}. In case \code{element.type = "ltr_element"} is chosen then the width of the left LTR element is used to visualize width vs. seq. similarity between the LTRs.}

\item{plot.type}{type of visualization: either \code{plot.type = "boxplot"} or \code{plot.type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. See Description for details.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{xlab}{label of x-axis.}

\item{ylab}{label of y-axis.}

\item{main}{title label.}

\item{legend.title}{title of legend.}
}
\description{
Visualize the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element.
}
\details{
This function visualizes the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element. Using this visualization approach, different classes of LTR retrotransposons can be detected due to their width and age correlation.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred(genome.file = "TAIR10_chr_all.fas",
                    trnas       = "plantRNA_Arabidopsis.fsa",
                    hmms        = "hmm_*")
                    
# visualize the collelation between LTR transposon age and width
# of predicted  A. thaliana LTR retrotransposons
PlotLTRWidth(Ath.Pred, plot.type = "violin")

# visualize the collelation between LTR retrotransposon age and width
# of predicted  A. thaliana LTR element
PlotLTRWidth(Ath.Pred, element.type = "ltr_element", plot.type = "violin")
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRharvest}}, \code{\link{LTRdigest}},
\code{\link{PlotLTRAge}}
}

