% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRWidthDistribution.R
\name{PlotLTRWidthDistribution}
\alias{PlotLTRWidthDistribution}
\title{Plot the LTR width distribution}
\usage{
PlotLTRWidthDistribution(LTRpred.tbl, type = "boxplot",
  similarity.bin = NULL, min.sim = NULL)
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{type}{type of visualization: either \code{type = "boxplot"} or \code{type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}
}
\description{
Function to visualize the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR element.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred()
# visualize the collelation between LTR transposon age and width
# of predicted  A. thaliana LTR element
PlotLTRWidthDistribution(Ath.Pred)
}

}
\author{
Hajk-Georg Drost
}

