% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinDomain.R
\name{PlotProteinDomain}
\alias{PlotProteinDomain}
\title{Visualize the Protein Domain distribution of predicted LTR retrotransposons}
\usage{
PlotProteinDomain(LTRpred.tbl, xlab = "\\nDetected Protein Domains",
  ylab = "Predicted LTR retrotransposons (count)\\n",
  main = "Protein Domain distribution of predicted LTR retrotransposons",
  legend.title = "Protein Domains")
}
\arguments{
\item{LTRpred.tbl}{\code{data.frame} returned by \code{\link{LTRpred}}.}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{main}{title text.}

\item{legend.title}{text of the legend title.
Either \code{annotation = "Dfam"} or \code{annotation = "Repbase"}.}
}
\details{
Plot the Superfamily distribution of predicted LTR retrotransposons of
annotations generated via Dfam or Repbase query.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)

# visualize superfamily distribution
PlotFamily(pred)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{read.ltrpred}}, \code{\link{dfam.query}}, \code{\link{repbase.query}}
}

