% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRanges.R
\name{PlotRanges}
\alias{PlotRanges}
\title{Plot Ranges of an genomic feature}
\usage{
PlotRanges(x, xlim = x, main = deparse(substitute(x)), col = "black",
  sep = 0.5, ...)
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

