% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRangesLTRharvestPrediction.R
\name{PlotRangesLTRharvestPrediction}
\alias{PlotRangesLTRharvestPrediction}
\title{Plot the location of putative LTR transposons along the chromosomes}
\usage{
PlotRangesLTRharvestPrediction(ltr.harvest.prediction)
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

