% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRepeatSpace.R
\name{plotRepeatSpace}
\alias{plotRepeatSpace}
\title{Genome Size vs. Repeat Space}
\usage{
plotRepeatSpace(genome.matrix, cor.method = "spearman",
  label.organism = TRUE, smooth.method = "glm", se = TRUE,
  xlab = "Total repeats length in Mega [bp]",
  ylab = "Genome size in Mega [bp]", main = "", alpha = 0.3,
  arrow_lab = FALSE, cl.analysis = FALSE, cl.centers = NULL,
  cl.method = "euclidean", cl.nstart = 100, cl.iter.max = 10000,
  sim.matrix = NULL, text.size = 18, label.size = 3,
  check.overlap = TRUE)
}
\arguments{
\item{genome.matrix}{Genome matrix retruned by \code{\link[LTRpred]{LTRpred.meta}}.}

\item{cor.method}{method to correlate genome size versus protein space.}

\item{label.organism}{shall organism names be drawn in the figure?}

\item{smooth.method}{method to add a smoothed conditional mean (see \code{ggplot2::geom_smooth()}).
Users can choose from \code{auto}, \code{glm}, \code{lm}, or \code{loess}. If no smoothing function shall be applied,
users can specify \code{smooth.method = 'NULL'}.}

\item{se}{shall standard error of the smoothing model (grey area) be drawn or not.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{alpha}{transparency of confidence interval of the smoothing function: between \[ 0,1 \].}

\item{arrow_lab}{whether or not data points shall be labelled with arrows (ggrepel).}

\item{cl.analysis}{logical value indicating whether or not cluster analysis of the \code{sim.matrix} returned by \code{\link[LTRpred]{LTRpred.meta}} shall be performed (see Details sections).}

\item{cl.centers}{number of expected clusters or a set of initial (distinct) cluster centres.}

\item{cl.method}{distance measure to perform cluster analysis. Options are \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, \code{"pearson"} , \code{"abspearson"} , \code{"abscorrelation"}, \code{"correlation"}, \code{"spearman"} or \code{"kendall"}.}

\item{cl.nstart}{If \code{cl.centers} is a number, number of random sets that shall be chosen.}

\item{cl.iter.max}{maximum number of iterations for cluster analysis.}

\item{sim.matrix}{Similarity matrix retruned by \code{\link[LTRpred]{LTRpred.meta}}.}

\item{text.size}{size of the labels in ggplot2 notation.}

\item{label.size}{size of the dot-labels.}

\item{check.overlap}{shall overlaps of labels be avoided or not.}
}
\description{
Compute and visualize the correlation between genome size vs. Repeat Space.
}
\details{
In the first step, users should have generated LTR retrotransposon annotations for several species using the
\code{\link[LTRpred]{LTRpred.meta}} function. The output of \code{\link[LTRpred]{LTRpred.meta}} is a folder storing the annotation files
for all species of input species. In addition, \code{\link[LTRpred]{LTRpred.meta}} generates two files: \code{*_SimilarityMatrix} and \code{*_GenomeInfo} .
}
\examples{
GenomeMatrix <- read.csv(system.file("GenomeMatrix.csv",package = "LTRpred"), sep = ";")
GenomeMatrix[ , 1] <- sapply(GenomeMatrix[ , 1],
                             function(x) unlist(stringr::str_split(x,"_"))[1])

plotSize(GenomeMatrix)
}
\author{
Hajk-Georg Drost
}
