% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSimCount.R
\name{PlotSimCount}
\alias{PlotSimCount}
\title{Plot LTR Similarity vs. predicted LTR count}
\usage{
PlotSimCount(sim.matrix, xlab = "LTR \% Similarity", ylab = "Count",
  main = "LTR \% Similarity vs. Count", text.size = 18)
}
\arguments{
\item{sim.matrix}{Similarity matrix retruned by \code{\link{LTRpred.meta}}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{text.size}{size of the labels in ggplot2 notation.}
}
\description{
The LTR Similarity vs. predicted LTR count visualization
allows to study the variability of predicted LTRs in different genomes.
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}

