% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.R
\name{plotSize}
\alias{plotSize}
\title{Genome size vs. LTR transposon content}
\usage{
plotSize(
  genome.summary,
  type = "total_ltrs_nucl_mbp",
  cor.method = "spearman",
  label.organism = TRUE,
  colour = "black",
  smooth.method = "glm",
  se = TRUE,
  xlab = "LTR retrotransposon content in Mega [bp]",
  ylab = "Genome size in Mega [bp]",
  main = "",
  alpha = 0.3,
  arrow_lab = FALSE,
  min.total.ltrs = 15,
  text.size = 18,
  label.size = 3,
  point.size = 3.5,
  check.overlap = TRUE,
  remove.zero = FALSE
)
}
\arguments{
\item{genome.summary}{a list generated by \code{\link{quality.filter.meta}} storing the \code{sim_file} and \code{gm_file} data.frames.}

\item{type}{type of LTR abundance. Options are: \code{type = "mass"} (total length of all TEs in Mbp), \code{type = "prop.mass"} (proportion of TEs within entire genome in %),
\code{type = "count"} (total number of TEs in genome), \code{type = "norm.count"} (total number of TEs in genome normalized by genome size in Mbp).}

\item{cor.method}{correlation analysis method. Either \code{"pearson"}, \code{"kendall"}, or \code{"spearman"}.}

\item{label.organism}{shall organism names be drawn in the figure?}

\item{colour}{a colour for the graph.}

\item{smooth.method}{method to add a smoothed conditional mean (see \code{ggplot2::geom_smooth()}).
Users can choose from \code{auto}, \code{glm}, \code{lm}, or \code{loess}. If no smoothing function shall be applied,
users can specify \code{smooth.method = 'NULL'}.}

\item{se}{shall standard error of the smoothing model (grey area) be drawn or not.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{alpha}{transparency of confidence interval of the smoothing function: between \[ 0,1 \].}

\item{arrow_lab}{whether or not data points shall be labelled with arrows (ggrepel).}

\item{min.total.ltrs}{minimum number of LTR retrotransposon abundance in Mbp to label dots.}

\item{text.size}{size of the labels in ggplot2 notation.}

\item{label.size}{size of the dot-labels.}

\item{point.size}{size of geom_points.}

\item{check.overlap}{shall overlaps of labels be avoided or not.}

\item{remove.zero}{shall species with 0 retrotransposons be removed?}
}
\description{
Compute and visualize the correlation between genome size vs. LTR transposon content.
}
\details{
In the first step, users should have generated LTR retrotransposon annotations for several species using the
\code{\link[LTRpred]{LTRpred.meta}} function. The output of \code{\link[LTRpred]{LTRpred.meta}} is a folder storing the annotation files
for all species of input species. In addition, \code{\link[LTRpred]{LTRpred.meta}} generates two files: \code{*_SimilarityMatrix} and \code{*_GenomeInfo} .
}
\examples{
GenomeMatrix <- read.csv(system.file("GenomeMatrix.csv",package = "LTRpred"), sep = ";")
GenomeMatrix[ , 1] <- sapply(GenomeMatrix[ , 1], 
                             function(x) unlist(stringr::str_split(x,"_"))[1])
                             
plotSize(GenomeMatrix)
}
\author{
Hajk-Georg Drost
}
