% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSizeCorrelation.R
\name{PlotSizeCorrelation}
\alias{PlotSizeCorrelation}
\title{Plot Genome size vs. LTR transposon count}
\usage{
PlotSizeCorrelation(genome.matrix, xlab = "LTR transposon Count",
  ylim = "Genome size in bp", main = "Genome size vs. LTR transp. count",
  text.size = 18)
}
\arguments{
\item{genome.matrix}{Genome matrix retruned by \code{\link{LTRpred.meta}}.}

\item{xlab}{x-axis label.}

\item{main}{main text.}

\item{text.size}{size of the labels in ggplot2 notation.}

\item{ylab}{y-axis label.}
}
\description{
Genome size vs. LTR transposon count.
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}

