% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeMulti.R
\name{plotSizeMulti}
\alias{plotSizeMulti}
\title{plotSize for \code{generate.multi.quality.filter.meta} output}
\usage{
plotSizeMulti(
  meta.summary.file,
  cor.method = "spearman",
  type = "total_ltrs_nucl_mbp",
  label.organism = TRUE,
  smooth.method = "lm",
  se = FALSE,
  main = "",
  xlab = "LTR retrotransposon content in Mega [bp]",
  ylab = "Genome size in Mega [bp]",
  alpha = 0.3,
  arrow_lab = FALSE,
  text.size = 18,
  label.size = 4,
  check.overlap = TRUE
)
}
\arguments{
\item{meta.summary.file}{a meta.summary.file file.}

\item{cor.method}{correlation method.}

\item{type}{type of TE load metric that shall be visualized. Options are
\itemize{
\item \code{type = "total_ltrs_nucl_mbp"}
\item \code{type = "n_ltrs_freq"}
\item \code{type = "n_ltrs"}
\item \code{type = "total_ltrs_nucl_freq"}
}}

\item{label.organism}{shall organism names be labeled next to each dot in the correlation plot?}

\item{smooth.method}{shall a smoothing function be applied to the correlation plot?}

\item{se}{shall standard.error be drawn when \code{smooth.method = TRUE}.}

\item{main}{title text.}

\item{xlab}{x-axis text.}

\item{ylab}{y-axis text.}

\item{alpha}{alpha of standard error band.}

\item{arrow_lab}{shall arrows be drawn between dots and label names?}

\item{text.size}{general text size.}

\item{label.size}{label text size.}

\item{check.overlap}{apply overlap correction for label names.}
}
\description{
Visualize the correlation between genome size and TE load for \code{\link{generate.multi.quality.filter.meta}} output.
}
\author{
Hajk-Georg Drost
}
