% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2bed.R
\name{pred2bed}
\alias{pred2bed}
\title{Format LTR prediction data to BED file format}
\usage{
pred2bed(LTR.data)
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.}
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}
to a \code{\link{data.frame}} in \code{BED} file format.
}
\details{
The BED file format is defined by: chromosome; start; end; columns.
}
\examples{
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff",package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv",package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# generate BED file
bed.file <- pred2bed(LTRfile$ltr.retrotransposon)
head(bed.file)

# write BED file to hard drive
write.table(bed.file,"test.bed", quote = FALSE, col.names = TRUE, row.names = FALSE)
}
\author{
Hajk-Georg Drost
}
\references{
http://www.ensembl.org/info/website/upload/bed.html
}

