% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2chromoMap.R
\name{pred2chromoMap}
\alias{pred2chromoMap}
\title{Format LTR prediction data to chromoMap package annotation file format}
\usage{
pred2chromoMap(LTR.data, output = "output")
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{output}{filename of the output chromoMap file.}
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRpred}} to a \code{chromoMap} package annotation file.
}
\details{
The chromoMap file format is defined by: element_name; chromosome_name; element_start; element_end; data.
}
\references{
https://cran.r-project.org/web/packages/chromoMap/vignettes/chromoMap.html
}
\author{
Hajk-Georg Drost
}
