% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2gff.R
\name{pred2gff}
\alias{pred2gff}
\title{Format LTR prediction data to GFF3 file format}
\usage{
pred2gff(LTR.data, output = "output.gff", program = "LTRpred")
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.}

\item{output}{filename of the output GFF file.}

\item{program}{program used to generate the prediction table, e.g. \code{program = "LTRpred"},
\code{program = "LTRdigest"}, or \code{program = "LTRharvest"}.}
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}
to a \code{\link{data.frame}} in \code{GFF3} file format.
}
\details{
The GFF3 file format is defined by: chromosome; start; end; name; columns.
}
\examples{
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff",
                        package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv"
                           ,package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# generate GFF file
pred2gff(LTRfile$ltr.retrotransposon, output = "test.gff")

}
\references{
http://www.ensembl.org/info/website/upload/gff.html
}
\author{
Hajk-Georg Drost
}
