% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2GRanges.R
\name{pred2GRanges}
\alias{pred2GRanges}
\title{Format LTR prediction data to \code{GRages} object}
\usage{
pred2GRanges(LTR.data, similarity.threshold = 95)
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{similarity.threshold}{the LTR similarity threshold that shall be used to define young and old retrotransposons.}
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRpred}} to a \code{GRages} object.
}
\details{
The \code{GRages} object is defined by the following columns:
\itemize{
\item \code{chromosome}
\item \code{start}
\item \code{end}
\item \code{strand}
\item \code{element_name}
\item \code{ltr_similarity}
\item \code{orfs}
\item \code{age}
}
}
\author{
Hajk-Georg Drost
}
