% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2orfseqs.R
\name{pred2orfseqs}
\alias{pred2orfseqs}
\title{Save the ORF sequence of the predicted LTR Transposons in a fasta file}
\usage{
pred2orfseqs(LTRpred.tbl, orf.seq.file, output = "output.fa")
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{orf.seq.file}{the fasta file storing the open reading frame sequences of predicted retroelements.
as returned by the \code{\link{LTRpred}} function.}

\item{output}{the fasta file to which the output sequences shall be stored in.}
}
\description{
This function allows users to save the sequence of the predicted LTR Transposons or LTRs in a fasta file.
}
\details{
The output \code{data.frame}s returned by \code{\link{LTRpred}} contain all information of the predicted
LTR retrotransposons that can be used for post-filtering steps. After these post-filtering steps
sequences of the remaining (filtered) candidates can be retrieved by this function.
}
\seealso{
\code{\link{LTRharvest}}, \code{\link{LTRdigest}}, \code{\link{LTRpred}}, \code{\link{read.prediction}}
}
\author{
Hajk-Georg Drost
}
