% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2tsv.R
\name{pred2tsv}
\alias{pred2tsv}
\title{Format LTR prediction data to TSV file format}
\usage{
pred2tsv(LTR.data, output = "output.tsv")
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.}

\item{output}{filename of the output CSV file.}
}
\value{
the \code{LTR.data} table in csv format saved to the hard drive.
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}
to a \code{\link{data.frame}} in \code{CSV} file format.
}
\examples{
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff",
                        package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv"
                           ,package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# generate csv file
pred2csv(LTRfile$ltr.retrotransposon)
}
\author{
Hajk-Georg Drost
}

