% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictORFs.R
\name{PredictORFs}
\alias{PredictORFs}
\title{Open Reading Frame Prediction}
\usage{
PredictORFs(input.file, orf.style = 7, min.codons = 200,
  trans.seqs = FALSE, output = NULL)
}
\description{
This function provides a wrapper to the \code{USEARCH}
\code{fastx_findorfs} command line tool to predict ORFs in
a given input fasta file and read the output as \code{\link{data.frame}} object.
}
\examples{
\dontrun{

}

}
\author{
Hajk-Georg Drost
}

