% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.filter.meta.R
\name{quality.filter.meta}
\alias{quality.filter.meta}
\title{Pipeline to eliminate false positive predictions of retrotransposons at a metagenomic scale}
\usage{
quality.filter.meta(kingdom, genome.folder, ltrpred.meta.folder, sim,
  cut.range = 2, n.orfs, strategy, update = FALSE)
}
\arguments{
\item{kingdom}{a character string specifying the kingdom of life to which genomes annotated with \code{\link{LTRpred.meta}} belong to. E.g. \code{kingdom = "Plants"}. If annotates of a variety of kingdoms have been done users can for example specify \code{kingdom = "Various"}.}

\item{genome.folder}{path to folder storing the genome assembly files that were used for \code{\link{LTRpred.meta}} predictions.}

\item{ltrpred.meta.folder}{path to folder storing the \code{\link{LTRpred.meta}} output files.}

\item{sim}{LTR similarity threshold. Only putative LTR transposons that fulfill this 
LTR similarity threshold will be retained.}

\item{cut.range}{a numeric number indicating the interval size for binning LTR similarities.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{strategy}{quality filter strategy. Options are
\itemize{
\item \code{strategy = "default"} : see section \code{Quality Control} 
\item \code{strategy = "stringent"} : in addition to filter criteria specified in section \code{Quality Control},
the filter criteria \code{!is.na(protein_domain)) | (dfam_target_name != "unknown")} is applied
}}

\item{update}{shall already existing \code{_SimilarityMatrix.csv} and \code{_GenomeInfo.csv} files be updated (\code{update = TRUE}) or can the already existing files be used (\code{update = FALSE})?}
}
\value{
A list with to list elements \code{sim_file} and \code{gm_file}. Each list element stores a \code{data.frame}:
  \itemize{
  \item \code{sim_file} (similarity file)
         \itemize{
         This \code{data.frame} stores the information 
                  \item 
                  }
   \item \code{gm_file} (genome metrics file)
         \itemize{
         This \code{data.frame} stores the information
                  \item
                  }
   }
}
\description{
This function takes the file paths to the genomes folder and \code{\link{LTRpred.meta}} output folder as input and eliminates false positive retrotransposon predictions on a metagenomic scale.
}
\details{
\strong{Quality Control}

\itemize{
\item \code{ltr.similarity}: Minimum similarity between LTRs. All TEs not matching this
 criteria are discarded.
 \item \code{n.orfs}: minimum number of Open Reading Frames that must be found between the
  LTRs. All TEs not matching this criteria are discarded.
 \item \code{PBS or Protein Match}: elements must either have a predicted Primer Binding
 Site or a protein match of at least one protein (Gag, Pol, Rve, ...) between their LTRs. All TEs not matching this criteria are discarded.
 \item The relative number of N's (= nucleotide not known) in TE <= 0.1. The relative number of N's is computed as follows: absolute number of N's in TE / width of TE.
}
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRpred.meta}}, \code{\link{read.ltrpred}}
}
\author{
Hajk-Georg Drost
}
