% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.filter.R
\name{quality.filter}
\alias{quality.filter}
\title{Try to eliminate false positive LTRpred predictions by filter strategy}
\usage{
quality.filter(pred, sim, n.orfs)
}
\arguments{
\item{pred}{\code{LTRpred.tbl} generated with \code{\link{LTRpred}}}

\item{sim}{LTR similarity threshold. Only putative LTR transposons that fulfill this 
LTR similarity threshold will be retained.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}
}
\value{
A quality filtered \code{LTRpred.tbl}.
}
\description{
This function takes an \code{\link{LTRpred}} output table as input
and eliminates false positive predictions.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
# read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)
# apply quality filter
pred <- quality.filter(pred, sim = 70, n.orfs = 1)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRpred.meta}}, \code{\link{read.ltrpred}}
}

