% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.cds.space.R
\name{quant.cds.space}
\alias{quant.cds.space}
\title{Quantify the coding sequence space within a genome}
\usage{
quant.cds.space(file)
}
\arguments{
\item{file}{file path to a fasta file storing the cds sequences.}
}
\description{
Quantification of the cds space (= total length of
all annotated coding sequences) within a given genome of interest.
}
\seealso{
\code{\link{quant.protein.space}}, \code{\link{quant.repeat.space}},
\code{\link{quant.meta.space}}
}
\author{
Hajk-Georg Drost
}
