% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.protein.space.R
\name{quant.protein.space}
\alias{quant.protein.space}
\title{Quantify the protein space within a genome}
\usage{
quant.protein.space(file)
}
\arguments{
\item{file}{file path to a fasta file storing the protein sequences.}
}
\description{
Quantification of the protein space (= total length of
all annotated protein coding genes) within a given genome of interest.
}
\seealso{
\code{\link{quant.cds.space}}, \code{\link{quant.repeat.space}},
\code{\link{quant.meta.space}}
}
\author{
Hajk-Georg Drost
}
