% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.repeat.space.R
\name{quant.repeat.space}
\alias{quant.repeat.space}
\title{Quantify the repeat space within a genome}
\usage{
quant.repeat.space(file)
}
\arguments{
\item{file}{file path to a fasta file storing the Repeat Masker annotation file.}
}
\description{
Quantification of the repeat space (= total length of
all Repeat Masker annotated repeats) within a given genome of interest.
}
\seealso{
\code{\link{quant.protein.space}}, \code{\link{quant.cds.space}}
}
\author{
Hajk-Georg Drost
}
