% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.space.meta.R
\name{quant.space.meta}
\alias{quant.space.meta}
\title{Quantify the genomic loci space for multiple genomes, proteomes or Repeat Masker annotation files}
\usage{
quant.space.meta(folder, type = "proteome")
}
\arguments{
\item{folder}{path to the folder storing the genomes, proteomes, or 
Repeat Masker files of interest.}

\item{type}{The following genomic features can be quantified:
\itemize{
\item \code{type = 'cds'}: coding sequence files in fasta format (see \code{\link[biomartr]{read_genome}}).
\item \code{type = 'proteome'}: proteome files in fasta format (see \code{\link[biomartr]{read_proteome}}).
\item \code{type = 'rm'}: Repeat Masker annotation files in fasta format (see \code{\link[biomartr]{read_rm}}).
}}
}
\description{
Quantification of the genomic loci space (= total length of
all annotated genomic features) within a given genome of interest.
}
\author{
Hajk-Georg Drost
}
