% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dfam.R
\name{read.dfam}
\alias{read.dfam}
\title{Import Dfam Query Output}
\usage{
read.dfam(dfam.file)
}
\arguments{
\item{dfam.file}{path to Dfam output file generated with \code{\link{dfam.query}}.}
}
\description{
Import the output file generated with \code{\link{dfam.query}}.
}
\examples{
# import example Dfam output
dfam.file <- read.dfam(system.file("example_dfam.out", package = "LTRpred"))

head(dfam.file)
}
\author{
Hajk-Georg Drost
}

