% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.findorfs.output.R
\name{read.findorfs.output}
\alias{read.findorfs.output}
\title{Read output of USEARCH's findorfs output}
\usage{
read.findorfs.output(findorfs.file)
}
\description{
This function reads the output of the \code{USEARCH}
\code{fastx_findorfs} command line tool and stores the 
sequence id and number of predicted ORFs in a \code{\link{data.frame}} object.
}
\author{
Hajk-Georg Drost
}

