% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.orfs.R
\name{read.orfs}
\alias{read.orfs}
\title{Read output of \code{ORFpred}}
\usage{
read.orfs(input.file)
}
\arguments{
\item{input.file}{fasta file generated by \code{\link{ORFpred}}.}
}
\value{
A \code{\link{data.frame}} object storing the \code{seq.id}, \code{orfs} (number of predicted ORFs), \code{start}, and \code{end} of the predicted LTRs.
}
\description{
This function reads the output of the \code{\link{ORFpred}} function and stores the 
sequence id and number of predicted ORFs in a \code{\link{data.frame}} object.
}
\details{
The file generated by \code{\link{ORFpred}} is parsed by this function
and returned as \code{\link{data.frame}} object.
}
\examples{
# read an example prediction file generated by PredictORFs()
ORFPred <- read.orfs(system.file("nt.fa",package = "LTRpred"))

head(ORFPred)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{ORFpred}}, \code{\link{LTRpred}}
}

