% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.prediction.R
\name{read.prediction}
\alias{read.prediction}
\title{Import the output of LTRharvest or LTRdigest}
\usage{
read.prediction(gff.file = NULL, tabout.file, program = "LTRharvest",
  ltr.fasta = NULL, inner.seq.fasta = NULL, data = NULL,
  similarity.bin = 2, min.sim = NULL)
}
\description{
This function
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}

