% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.seqs.R
\name{read.seqs}
\alias{read.seqs}
\title{Import sequences of predicted LTR transposons}
\usage{
read.seqs(seq.file, program = "LTRharvest")
}
\arguments{
\item{seq.file}{path to fasta file storing the sequences of predicted LTR transposons generated by 
\code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.}

\item{program}{program used to generate the LTR transposons specified in \code{seq.file}, e.g. \code{program = "LTRpred"}, \code{program = "LTRdigest"}, or \code{program = "LTRharvest"}.}
}
\description{
Import sequences of predicted LTR transposons generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.
}
\author{
Hajk-Georg Drost
}
