% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tabout.R
\name{read.tabout}
\alias{read.tabout}
\title{Import information sheet returned by LTRdigest}
\usage{
read.tabout(tabout.file)
}
\arguments{
\item{tabout.file}{path to the tabout.csv file generated by the \code{\link{LTRdigest}} function.}
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

