% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.uc.R
\name{read.uc}
\alias{read.uc}
\title{Read file in USEARCH cluster format}
\usage{
read.uc(uc.file)
}
\arguments{
\item{uc.file}{path to file in USEARCH cluster format (\code{*.uc} file extension).}
}
\value{
A dataframe storing the following columns:

\itemize{
\item \code{Type:} Record type 'S', 'H', 'C', or 'N'.
\item \code{Cluster:} Cluster number (0-based).
\item \code{Size:} Sequence length ('S', 'N', and 'H') or cluster size 'C'.
\item \code{Perc_Ident:} For 'H' records, percent identity with target.
\item \code{Strand:} For 'H' records, the strand: '+' or '-' for nucleotides; '.' for proteins.
\item \code{Query:}  query id.
\item \code{Target:} target id.
}

Details:

Record type: 
\itemize{
\item \code{Type 'H' :} Hit. Represents an alignment between the query sequence and target sequence. For clustering 'H' indicates the cluster assignment for the query.
\item \code{Type 'S' :} Centroid (clustering only). There exists only one 'S' record
for each cluster, this gives the centroid (representative) sequence label in the \code{Query} column.
\item \code{Type 'C' :} Cluster record (clustering only). The \code{Size} column specifies the cluster size and the \code{Query} column the query id that corresponds to this cluster.
\item \code{Type 'N' :} No hit (for database search without clustering only). Indicates that no hit of the query were found in the target database. In the case of clustering, a query without hits becomes the centroid of a new cluster and generates an 'S'
record instead of an 'N' record.
}
}
\description{
Read a file in USEARCH cluster format generated by either USEARCH or VSEARCH.
}
\examples{
# read example *.uc file
test.uc <- read.uc(system.file("test.uc", package = "LTRpred"))

# look at the format in R
head(test.uc)
}
\author{
Hajk-Georg Drost
}

