% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLTRharvestPredictionSeqs.R
\name{ReadLTRharvestPredictionSeqs}
\alias{ReadLTRharvestPredictionSeqs}
\title{Import sequences of LTRharvest predicted LTR transposons}
\usage{
ReadLTRharvestPredictionSeqs(ltr.fasta.file)
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

