% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.fasta.meta.R
\name{rename.fasta.meta}
\alias{rename.fasta.meta}
\title{Meta function for applying \code{rename.fasta}}
\usage{
rename.fasta.meta(in.folder, out.file)
}
\arguments{
\item{in.folder}{path to folder storing \code{\link{LTRpred}} generated organism prediction folders.}

\item{out.file}{name of concatenated output file.}
}
\description{
Meta function for applying \code{\link{rename.fasta}} to many fasta files within a folder.
}
\author{
Hajk-Georg Drost
}
