% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.fasta.R
\name{rename.fasta}
\alias{rename.fasta}
\title{Helper function to add species names to headers within fasta files}
\usage{
rename.fasta(file, species, output = "renamed_fasta.fa", append = FALSE)
}
\arguments{
\item{file}{input fasta file.}

\item{species}{a character string specifying the species name to be added to the headers. 
Format will be: \code{species_}*, where * stands for the original header.}

\item{output}{a character string denoting the name of the renamed output fasta file.}

\item{append}{logical value. If \code{TRUE} output will be appended to file; otherwise, it will overwrite the contents of file.}
}
\value{
Writes a new fasta file with renamed headers.
}
\description{
Reads a fasta file and adds the defined \code{species} name to each
header entry of the input fasta file.
}
\examples{
# example fasta file
seqs <- system.file("nt.fa",package = "LTRpred")
# rename headers in file by adding species name: "Athaliana"
rename.fasta(file = seqs, species = "Athaliana", output = "Athaliana.fa")
}
\author{
Hajk-Georg Drost
}

