% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.organisms.R
\name{rename.organisms}
\alias{rename.organisms}
\title{Rename file path names from \code{read_proteome}, etc.}
\usage{
rename.organisms(data)
}
\arguments{
\item{data}{a GenomeInfo file.}
}
\description{
Rename a file path e.g. \code{Arabidopsis_thaliana.fa.gz} to \code{Athaliana}.
}
\author{
Hajk-Georg Drost
}
