% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repbase.clean.R
\name{repbase.clean}
\alias{repbase.clean}
\title{Clean the initial Repbase database for BLAST}
\usage{
repbase.clean(repbase.file, output.file)
}
\arguments{
\item{repbase.file}{fasta file storing the corresponding Repbase annotation, e.g. \code{athrep.ref}.}

\item{output.file}{name/path of the cleaned Repbase annotation file.}
}
\description{
Clean the headers of the Repbase fasta files
so that headers can be used to create a blast-able database.
}
\details{
The Repbase database can be downloaded after registration at http://www.girinst.org/repbase/.
The corresponding files as they are however, cannot be converted into a blast-able database.
Hence, a pre-filtering step is neccessary to be able to use this database with the e.g. \code{\link{repbase.query}}
function.
}
\examples{
\dontrun{
# PreProcess Repbase: A thaliana
# and save the output into the file "Athaliana_repbase.ref"
repbase.clean(repbase.file = "athrep.ref",
              output.file  = "Athaliana_repbase.ref")

}
}
\references{
http://www.girinst.org/repbase/
}
\author{
Hajk-Georg Drost
}
