% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimMatAbundance.R
\name{SimMatAbundance}
\alias{SimMatAbundance}
\title{Compute histogram shape similarity between species}
\usage{
SimMatAbundance(sim.matrix)
}
\arguments{
\item{sim.matrix}{a species age distribution matrix generated with \code{\link{LTRpred.meta}}.}
}
\description{
For each pairwise species comparison the pattern similarity between
LTR age distributions is computed via \code{1 - cor(species1,species2)}.
}
\author{
Hajk-Georg Drost
}
