% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te.cn.R
\name{te.cn}
\alias{te.cn}
\title{Copy Number Quantification of predicted LTR transposons}
\usage{
te.cn(LTRpred.folder, genome, ltr.similarity = 70, scope.cutoff = 0.85,
  perc.ident.cutoff = 70, output = NULL, max.hits = 65000, eval = 1e-10)
}
\arguments{
\item{LTRpred.folder}{file path to the \code{\link{LTRpred}} output folder.}

\item{genome}{file path to the reference genome in which solo LTRs shall be found (in \code{fasta} format).}

\item{ltr.similarity}{similarity threshold for defining LTR similarity.}

\item{output}{file name of the BLAST output. If \code{output = NULL} (default) then the BLAST output file will be deleted after the result \code{data.frame} is returned by this function.}

\item{max.hits}{maximum number of hits that shall be retrieved that still fulfill the e-value criterium.
Default is \code{max.hits = 65000}.}

\item{eval}{e-value threshold for BLAST hit detection. Default is \code{eval = 1E-10}.}
}
\description{
Detect copies of predicted LTR transposons and report genomic locations of corresponding copies using a BLAST search strategy.
}
\details{
The
}
\examples{
\dontrun{

}
}
\author{
Hajk-Georg Drost
}
\references{
Altschul, S.F., Gish, W., Miller, W., Myers, E.W. & Lipman, D.J. (1990) "Basic local alignment search tool." J. Mol. Biol. 215:403-410.

Gish, W. & States, D.J. (1993) "Identification of protein coding regions by database similarity search." Nature Genet. 3:266-272.

Madden, T.L., Tatusov, R.L. & Zhang, J. (1996) "Applications of network BLAST server" Meth. Enzymol. 266:131-141.

Altschul, S.F., Madden, T.L., Schaeffer, A.A., Zhang, J., Zhang, Z., Miller, W. & Lipman, D.J. (1997) "Gapped BLAST and PSI-BLAST: a new generation of protein database search programs." Nucleic Acids Res. 25:3389-3402.

Zhang Z., Schwartz S., Wagner L., & Miller W. (2000), "A greedy algorithm for aligning DNA sequences" J Comput Biol 2000; 7(1-2):203-14.
}

