% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.datasheet.R
\name{tidy.datasheet}
\alias{tidy.datasheet}
\title{Select most important columns of \code{LTRpred} output for further analytics}
\usage{
tidy.datasheet(LTRpred.tbl)
}
\arguments{
\item{LTRpred.tbl}{a \code{data.frame} storing the result (DataSheet) of \code{\link{LTRpred}}.}
}
\description{
This function simply selects the most important columns of \code{LTRpred} output for further analytics.
}
\author{
Hajk-Georg Drost
}
