% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.wrap.path.R
\name{ws.wrap.path}
\alias{ws.wrap.path}
\title{Wrap whitespace in paths}
\usage{
ws.wrap.path(paths)
}
\arguments{
\item{paths}{a string or vector of strings containing paths.}
}
\description{
Helper function for wrapping white spaces in folder paths.
}
\author{
Hajk-Georg Drost
}

