% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.jumpers.R
\name{filter.jumpers}
\alias{filter.jumpers}
\title{Detect LTR retrotransposons that are potential jumpers}
\usage{
filter.jumpers(LTRpred.tbl, ltr.similarity = 95, strategy = "conservative")
}
\arguments{
\item{LTRpred.tbl}{a \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{ltr.similarity}{LTR similarity threshold. Default is \code{ltr_similarity = 95}.}

\item{strategy}{filter strategy: either \code{conservative}, \code{liberal}, or \code{between}.}
}
\description{
This function applies specific filter criteria to
screen for LTR retrotransposons predicted by \code{\link{LTRpred}}
that are potentially able to transpose due to their sequence features.
}
\details{
This ...

\strong{Filter strategy}
\itemize{
\item \code{conservative} :
\item \code{liberal} :
\item \code{between} :
}
}
\examples{
\dontrun{
# generate de novo LTR transposon prediction
pred <- LTRpred(genome.file = "TAIR10_chr_all.fas",
                trnas       = "plantRNA_Arabidopsis.fsa",
                hmms        = "hmm_*")
                
# detect potential jumpers               
filter.jumpers(pred)
}
}
\seealso{
\code{\link{LTRpred}}
}
\author{
Hajk-Georg Drost
}
