% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.summary.R
\name{genome.summary}
\alias{genome.summary}
\title{Generating genome summary files for \code{LTRpred.meta} results}
\usage{
genome.summary(genome.folder, ltrpred.meta.folder, file.name, sim = 70,
  cut.range = 2, quality.filter = TRUE, n.orfs = 0,
  strategy = "default")
}
\description{
Generating genome summary files for \code{LTRpred.meta} results
}
\details{
...
}
\author{
Hajk-Georg Drost
}
