% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.with.genome.tbl.R.R
\name{join.with.genome.tbl}
\alias{join.with.genome.tbl}
\title{Join \code{gm_files} returned by \code{generate.multi.quality.filter.meta} with a genome information table}
\usage{
join.with.genome.tbl(meta.summary.file, genome.tbl)
}
\description{

}
\author{
Hajk-Georg Drost
}
