% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRharvest.R
\name{LTRharvest}
\alias{LTRharvest}
\title{Run LTRharvest to predict putative LTR Retrotransposons}
\usage{
LTRharvest(input, index.file = NULL, seed = 30, minlenltr = 100,
  maxlenltr = 3500, mindistltr = 4000, maxdistltr = 25000, similar = 70,
  mintsd = 4, maxtsd = 20, vic = 60, overlaps = "no", xdrop = 5,
  mat = 2, mis = -2, ins = -3, del = -3, motifmis = 0, motif = NULL,
  output.path = NULL)
}
\description{
This function implements an interface between R and
the LTRharvest command line tool to predict putative LTR retrotransposons from R.
}
\details{

}
\examples{
\dontrun{

}

}
\author{
Hajk-Georg Drost
}

