% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRpred.meta.R
\name{LTRpred.meta}
\alias{LTRpred.meta}
\title{Perform Meta-Analyses with LTRpred}
\usage{
LTRpred.meta(genome.folder, output.name, cores = 1, ...)
}
\arguments{
\item{\dots}{}
}
\description{
Run \code{\link{LTRpred}} on several genomes (sequentially) that are stored in a given folder.
}
