% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTRpred.meta.R
\name{LTRpred.meta}
\alias{LTRpred.meta}
\title{Perform Meta-Analyses with LTRpred}
\usage{
LTRpred.meta(genome.folder = NULL, result.folder = NULL, sim = 70,
  cut.range = 2, quality.filter = TRUE, n.orfs = 1, file.name = NULL,
  LTRpred.meta.folder = NULL, ...)
}
\arguments{
\item{genome.folder}{path to the folder storing all unmasked genomes for which \code{\link{LTRpred}}
based \code{de novo} LTR retrotransposon prediction shall be performed.}

\item{result.folder}{folder in which \code{LTRpred} results shall be stored.}

\item{sim}{similarity threshold for defining LTR similarity.}

\item{cut.range}{range to bin similarity values (e.g.(98,100])).}

\item{quality.filter}{shall false positives be filtered out as much as possible or not. 
See \code{Description} for details.}

\item{n.orfs}{minimum number of Open Reading Frames that must be found between the LTRs (if \code{quality.filter = TRUE}). See \code{Details} for further information on quality control.}

\item{file.name}{name of the \code{*._SimilarityMatrix.csv} and \code{*.GenomeInfo.csv} file generated by this function.}

\item{LTRpred.meta.folder}{meta-folder storing already pre-cumputed \code{\link{LTRpred}} generated files.}

\item{\dots}{all parameters of \code{\link{LTRpred}}.}
}
\value{
\itemize{
\item a file \code{*_SimilarityMatrix.csv} is generated and stored in the working directory.
\item a file \code{*_GenomeInfo.csv} is generated and stored in the working directory.
}
}
\description{
Run \code{\link{LTRpred}} on several genomes (sequentially) that are stored in a given folder.
}
\details{
This function provides a crawler to run \code{\link{LTRpred}} sequencially
on any number of genomes stored within the same folder.
The result will be saved on in the \code{result.folder} and can be analysed with the
\code{\link{meta.apply}} function.

\strong{Quality Control}

\itemize{
\item \code{ltr.similarity}: Minimum similarity between LTRs. All TEs not matching this
 criteria are discarded.
 \item \code{n.orfs}: minimum number of Open Reading Frames that must be found between the
  LTRs. All TEs not matching this criteria are discarded.
 \item \code{PBS or Protein Match}: elements must either have a predicted Primer Binding
 Site or a protein match of at least one protein (Gag, Pol, Rve, ...) between their LTRs. All TEs not matching this criteria are discarded.
 \item The relative number of N's (= nucleotide not known) in TE <= 0.1. The relative number of N's is computed as follows: absolute number of N's in TE / width of TE.
}
}
\examples{
\dontrun{
# perform a meta analysis on multiple genomes
# stored in a genomes folder
LTRpred.meta(genome.folder = "Genomes/",
             result.folder = "LTRpredResults"
             trnas         = "plantRNA_Arabidopsis.fsa",
             hmms          = "hmm_*")
}
}
\author{
Hajk-Georg Drost
}

