% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot correlation between #cluster members and any other variable}
\usage{
plotCluster(
  data,
  y = "CG_3ltr_rel",
  color.by = "fam",
  quality.filter = FALSE,
  sim = 70,
  n.orfs = 1,
  chop.names = TRUE,
  xlab = "# CLUSTER MEMBERS",
  ylab = "",
  main = "",
  legend.title = "Protein Domain"
)
}
\arguments{
\item{data}{\code{data.frame} returned by \code{LTRpred}.}

\item{y}{column name in \code{data} (lazy evaluation) that shall be correlated with \code{cn_3ltr} (copy number of predicted 3' LTR).}

\item{color.by}{color points by a column name in \code{data} (lazy evaluation).}

\item{quality.filter}{shall a quality filter be applied to reduce false positives?}

\item{sim}{filter minimum sequence similarity between the LTRs.}

\item{n.orfs}{filter for minimum number of orfs.}

\item{chop.names}{shall superfamily names be chopped?}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{legend.title}{legend text.}
}
\description{
Visualize the correlation between the Cluster Number generated 
by \code{LTRpred} and any other variable 
and color each putative LTR transposon by it's e.g. superfamily name.
}
\author{
Hajk-Georg Drost
}
