% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRAgeDistribution.R
\name{PlotLTRAgeDistribution}
\alias{PlotLTRAgeDistribution}
\title{Plot the age distribution of predicted LTR transposons}
\usage{
PlotLTRAgeDistribution(LTRpred.tbl, similarity.bin = NULL, min.sim = NULL)
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}
}
\description{
This function visualizes the age distribution of
predicted LTR transposons generated with \code{\link{LTRpred}}.

The age pf LTR transposons is defined by the sequence similarity of their
LTRs. Evolutionary young (recent) LTR transposons tend to have very similar
LTRs (up to 100\% sequence similarity), whereas evolutionary older LTR transposons
tend to have less similar LTRs.
}
\details{
This way of visualizing the age distribution of LTR transposons
allows users to examine the rate of recent transposition events in extant organisms.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred()
# visualize the age distribution of predicted  A. thaliana LTR transposons
PlotLTRAgeDistribution(Ath.Pred)
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRharvest}}, \code{\link{LTRdigest}},
\code{\link{PlotLTRAgeDistribution}}, \code{\link{PlotLTRTransposonWidthDistribution}},
\code{\link{PlotLTRWidthDistribution}}, \code{\link{PlotRanges}}
}

