% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRRange.R
\name{PlotLTRRange}
\alias{PlotLTRRange}
\title{Plot the location of putative LTR transposons along the chromosomes}
\usage{
PlotLTRRange(LTRpred.tbl)
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}
}
\description{
Visualization of the location of putative LTR transposons along chromosomes.
}
\author{
Hajk-Georg Drost
}

