% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRTransposonWidthDistribution.R
\name{PlotLTRTransposonWidthDistribution}
\alias{PlotLTRTransposonWidthDistribution}
\title{Plot the width distribution of putative LTR transposons}
\usage{
PlotLTRTransposonWidthDistribution(LTRpred.tbl, type = "boxplot",
  similarity.bin = NULL, min.sim = NULL)
}
\arguments{
\item{LTRpred.tbl}{the \code{\link{data.frame}} generated by \code{\link{LTRpred}}.}

\item{type}{type of visualization: either \code{type = "boxplot"} or \code{type = "violin"}.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}
}
\description{
This
}
\details{
This function visualizes the correlation between LTR transposon age
(measured in LTR similarity) and the width of the corresponding LTR retrotransposon (the complete sequence). Using this approach, classes of LTR retrotransposons can be determined 
purely due to their width and age.
}
\examples{
\dontrun{
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred()
# visualize the collelation between LTR transposon age and width
# of predicted  A. thaliana LTR transposons
PlotLTRTransposonWidthDistribution(Ath.Pred)
}

}
\author{
Hajk-Georg Drost
}

