% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLTRWidthDistribution.R
\name{PlotLTRWidthDistribution}
\alias{PlotLTRWidthDistribution}
\title{Plot the LTR width distribution}
\usage{
PlotLTRWidthDistribution(ltr.harvest.prediction, type = "boxplot",
  similarity.bin = NULL, min.sim = NULL)
}
\details{

}
\examples{
\dontrun{

}

}
\author{
Hajk-Georg Drost
}

