% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMainInterSpeciesCluster.R
\name{plotMainInterSpeciesCluster}
\alias{plotMainInterSpeciesCluster}
\title{Plot inter species similarity between TEs (for the top 6 clusters)}
\usage{
plotMainInterSpeciesCluster(
  cluster.file,
  top.n = 6,
  decreasing = TRUE,
  xlab = "Species",
  ylab = "\% Similarity to Reference",
  legend.title = "Species"
)
}
\arguments{
\item{cluster.file}{a cluster file in \code{*.uc} format (imported with e.g. \code{\link[LTRpred]{filter.uc}} and \code{\link[LTRpred]{read.uc}}).}

\item{top.n}{number of top clustern that shall be visualized.}

\item{decreasing}{shall clusters be shown by size in decreasing order?}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{legend.title}{legend title text.}
}
\description{
After clustering TE sequences across different species with \code{\link[LTRpred]{CLUSTpred}}
the similarity of TEs from different species for the top 6 clusters is being visualized.
}
\author{
Hajk-Georg Drost
}
