% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRanges.R
\name{plotRanges}
\alias{plotRanges}
\title{Plot Ranges of an genomic feature}
\usage{
plotRanges(x, xlim = x, main = deparse(substitute(x)), col = "black",
  sep = 0.5, ...)
}
\arguments{
\item{x}{range object.}

\item{xlim}{range of the x-axis.}

\item{main}{main text for the plot generated.}

\item{col}{color of the visualized ranges.}

\item{sep}{separation between ranges.}

\item{...}{standard graphics parameters passed to \code{link{rect}}.}
}
\description{
Helper function to plot the genomic range of the predicted LTR transposon.
}
\author{
Hajk-Georg Drost
}
