% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSimCount.R
\name{PlotSimCount}
\alias{PlotSimCount}
\title{Plot LTR Similarity vs. predicted LTR count}
\usage{
PlotSimCount(sim.matrix, genome.matrix = NULL, type = "normalized",
  cl.analysis = FALSE, cl.centers = NULL, cl.method = "euclidean",
  cl.nstart = 100, cl.iter.max = 10000, min.sim = 70,
  similarity.bin = 2, xlab = "\% Similarity between 5' and 3' LTRs",
  ylab = "LTR retrotransposon content in Mega [bp]", main = "",
  text.size = 18)
}
\arguments{
\item{sim.matrix}{Similarity matrix retruned by \code{\link{LTRpred.meta}}.}

\item{genome.matrix}{Genome matrix retruned by \code{\link{LTRpred.meta}}.}

\item{type}{either absolute count value or normalized count value (normalized by genome size is shown): \code{type = "absolute"} or \code{type = "normalized"}. Default is \code{type = "absolute"}.}

\item{cl.analysis}{logical value indicating whether or not cluster analysis of the \code{sim.matrix} returned by \code{\link{LTRpred.meta}} shall be performed (see Details sections).}

\item{cl.centers}{number of expected clusters or a set of initial (distinct) cluster centres.}

\item{cl.method}{distance measure to perform cluster analysis. Options are \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, \code{"pearson"} , \code{"abspearson"} , \code{"abscorrelation"}, \code{"correlation"}, \code{"spearman"} or \code{"kendall"}.}

\item{cl.nstart}{If \code{cl.centers} is a number, number of random sets that shall be chosen.}

\item{cl.iter.max}{maximum number of iterations for cluster analysis.#' @param xlab x-axis label.}

\item{min.sim}{minimum similarity that was used with \code{\link{LTRpred.meta}}.}

\item{similarity.bin}{resolution of similarity binning that was used with \code{\link{LTRpred.meta}}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{main text.}

\item{text.size}{size of the labels in ggplot2 notation.}
}
\description{
The LTR Similarity vs. predicted LTR count visualization
allows to study the variability of predicted LTRs in different genomes.
}
\details{
In case the \code{sim.matrix} includes more than 30 observations (genomes), then
a \code{z.test} is applied to compute pairwise differences between neighboring distributions.

In case the \code{sim.matrix} includes less than 30 observations (genomes), then
a \code{wilcox.test} is applied to compute pairwise differences between neighboring distributions.
}
\examples{
SimMatrix <- read.csv(system.file("SimMatrix.csv",package = "LTRpred"), sep = ";")
GenomeMatrix <- read.csv(system.file("GenomeMatrix.csv",package = "LTRpred"), sep = ";")
names(SimMatrix) <- c("organism",levels(cut(100,rev(seq(100,70,-2)),
                                 include.lowest = TRUE,right = TRUE)))
SimMatrix2 <- SimMatrix
SimMatrix2[ , 2:16] <- SimMatrix2[ , 2:16] / GenomeMatrix$genome.size

PlotSimCount(SimMatrix)
PlotSimCount(SimMatrix2)
}
\author{
Hajk-Georg Drost
}

