% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2gff.R
\name{pred2gff}
\alias{pred2gff}
\title{Format LTR prediction data to GFF3 file format}
\usage{
pred2gff(LTR.data)
}
\arguments{
\item{LTR.data}{the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}.}
}
\description{
This function formats the LTR prediction \code{\link{data.frame}}
generated by \code{\link{LTRharvest}}, \code{\link{LTRdigest}}, or \code{\link{LTRpred}}
to a \code{\link{data.frame}} in \code{GFF3} file format.
}
\details{
The GFF3 file format is defined by: chromosome; start; end; name; columns.
}
\examples{
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff",
                        package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv"
                           ,package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# generate GFF file
gff.example <- pred2gff(LTRfile$ltr.retrotransposon)
head(gff.example)

# write GFF file to hard drive
write.table(gff.example,"test.gff", quote = FALSE, col.names = TRUE, row.names = FALSE)
}
\author{
Hajk-Georg Drost
}
\references{
http://www.ensembl.org/info/website/upload/gff.html
}

