% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.filter.R
\name{quality.filter}
\alias{quality.filter}
\title{Pipeline to eliminate false positive predictions of retrotransposons}
\usage{
quality.filter(pred, sim, n.orfs, strategy = "default")
}
\arguments{
\item{pred}{\code{LTRpred.tbl} generated with \code{\link{LTRpred}}}

\item{sim}{LTR similarity threshold. Only putative LTR transposons that fulfill this 
LTR similarity threshold will be retained.}

\item{n.orfs}{minimum number of ORFs detected in the putative LTR transposon.}

\item{strategy}{quality filter strategy. Options are
\itemize{
\item \code{strategy = "default"} : see section \code{Quality Control} 
\item \code{strategy = "stringent"} : in addition to filter criteria specified in section \code{Quality Control},
the filter criteria \code{!is.na(protein_domain)) | (dfam_target_name != "unknown")} is applied
}}
}
\value{
A quality filtered \code{LTRpred.tbl}.
}
\description{
This function takes an \code{\link{LTRpred}} output table as input
and eliminates false positive predictions.
}
\details{
\strong{Quality Control}

\itemize{
\item \code{ltr.similarity}: Minimum similarity between LTRs. All TEs not matching this
 criteria are discarded.
 \item \code{n.orfs}: minimum number of Open Reading Frames that must be found between the
  LTRs. All TEs not matching this criteria are discarded.
 \item \code{PBS or Protein Match}: elements must either have a predicted Primer Binding
 Site or a protein match of at least one protein (Gag, Pol, Rve, ...) between their LTRs. All TEs not matching this criteria are discarded.
 \item The relative number of N's (= nucleotide not known) in TE <= 0.1. The relative number of N's is computed as follows: absolute number of N's in TE / width of TE.
}
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
# read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)
# apply quality filter
pred <- quality.filter(pred, sim = 70, n.orfs = 1)
}
\seealso{
\code{\link{LTRpred}}, \code{\link{LTRpred.meta}}, \code{\link{read.ltrpred}}, \code{\link{quality.filter.meta}}
}
\author{
Hajk-Georg Drost
}
