% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ltrpred.R
\name{read.ltrpred}
\alias{read.ltrpred}
\title{Import the data sheet file generated by \code{LTRpred}}
\usage{
read.ltrpred(data.sheet)
}
\arguments{
\item{data.sheet}{path to the \code{*_LTRpred_DataSheet.csv} file.}
}
\description{
The \code{*_LTRpred_DataSheet.csv} file generated by \code{\link{LTRpred}} stores the featured of all predicted LTR transposons in a table. This function
imports this table.
}
\examples{
# example prediction file generated by LTRpred 
pred.file <- system.file("Athaliana_TAIR10_chr_all_LTRpred_DataSheet.csv", package = "LTRpred")
# read LTRpred generated prediction file (data sheet)
pred <- read.ltrpred(pred.file)

# or arrange by ltr_similarity
dplyr::arrange(tidy.datasheet(pred), dplyr::desc(ltr_similarity))
}
\author{
Hajk-Georg Drost
}

