% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.prediction.R
\name{read.prediction}
\alias{read.prediction}
\title{Import the output of LTRharvest, LTRdigest, or LTRpred}
\usage{
read.prediction(gff.file = NULL, tabout.file, program = "LTRdigest",
  ltr.fasta = NULL, inner.seq.fasta = NULL, data = NULL,
  similarity.bin = 2, min.sim = NULL)
}
\arguments{
\item{gff.file}{path to GFF file generated by either \code{\link{LTRharvest}}, \code{\link{LTRdigest}},
or \code{\link{LTRpred}}.}

\item{tabout.file}{path to \code{*_tabout.csv} file generated by either \code{\link{LTRdigest}} or
or \code{\link{LTRpred}}.}

\item{program}{prediction program used to generate the corresponding GFF file. 
Either \code{program = "LTRharvest"}, \code{program = "LTRdigest"}, or \code{program = "LTRpred"}.}

\item{ltr.fasta}{fasta file generated by either \code{\link{LTRdigest}} or
or \code{\link{LTRpred}} storing the sequences of the predicted LTR retrotransposons (e.g. \code{*-ltrdigest_complete.fas}).}

\item{inner.seq.fasta}{path to the \code{*_BetweenLTRSeqs.fsa} fasta file returned by \code{\link{LTRharvest}} in caste \code{program = "LTRharvest"} is specified.}

\item{data}{the \code{\link{data.frame}} returned by either \code{\link{LTRharvest}}, \code{\link{LTRdigest}},
or \code{\link{LTRpred}} that can be used to modify for example \code{similarity.bin} and \code{min.sim} within this table.}

\item{similarity.bin}{resolution of similarity binning. E.g. binning 98\%-100\% into 0.5\% intervals would be \code{similarity.bin = 0.5}.}

\item{min.sim}{minimum similarity between LTRs that can shall be considered for visualization.}
}
\description{
This function
}
\examples{
\dontrun{
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff", package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv" ,package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# look at results
str(LTRfile$ltr.retrotransposon)

}
}
\author{
Hajk-Georg Drost
}

