% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadLTRharvestPredictionSeqs.R
\name{read.seqs}
\alias{read.seqs}
\title{Import sequences of predicted LTR transposons}
\usage{
read.seqs(seq.file, program = "LTRharvest")
}
\arguments{
\item{seq.file}{a sequence file in fasta format.}
}
\description{
This function
}
\author{
Hajk-Georg Drost
}

