% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tabout.R
\name{read.tabout}
\alias{read.tabout}
\title{Import information sheet returned by LTRdigest}
\usage{
read.tabout(tabout.file)
}
\arguments{
\item{tabout.file}{path to the tabout.csv file generated by the
\code{\link{LTRdigest}} function.}
}
\description{
This function imports the \code{*-ltrdigest_tabout.csv} file generated by \code{\link{LTRdigest}}.
}
\examples{
# example tabout file generated by LTRdigest for A thaliana
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv",package = "LTRpred")
# import tabout file
imported.tabout <- read.tabout(tabout.file)
# look at the imported tabout file.
head(imported.tabout)      
}
\author{
Hajk-Georg Drost
}

