% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalRepeatSpace.R
\name{totalRepeatSpace}
\alias{totalRepeatSpace}
\title{Quantify the total repeat space from Repeat Masker output in Mbp}
\usage{
totalRepeatSpace(file, type = "LTR")
}
\arguments{
\item{file}{path to Repeat Masker output file.}

\item{type}{type of element for which total repeat space shall be quantified.
Options are:

\itemize{
 \item \code{type = "all"} include all types of repeats.
 \item \code{type = "LTR"} include only LTR retrotransposons.
}}
}
\author{
Hajk-Georg Drost
}
